extract dataset functions to a service
