+++ src/pages/DACDatasets.js
@@ -1,16 +1,13 @@
 import React, {useCallback, useEffect, useRef, useState} from 'react';
-import {DAC} from '../libs/ajax';
 import {Styles} from '../libs/theme';
 import lockIcon from '../images/lock-icon.png';
 import SearchBar from '../components/SearchBar';
 import {DACDatasetsTable, DACDatasetTableColumnOptions} from '../components/dac_dataset_table/DACDatasetsTable';
-import {getSearchFilterFunctions, Notifications, searchOnFilteredList, USER_ROLES} from '../libs/utils';
+import {getSearchFilterFunctions, Notifications, searchOnFilteredList} from '../libs/utils';
 import {consoleTypes} from '../components/dac_dataset_table/DACDatasetTableCellData';
 import style from './DACDatasets.module.css';
-import {Storage} from '../libs/storage';
-import {filter, flow, isNil, map} from 'lodash/fp';
 import {Button} from '@mui/material';
-import {getDataUseTranslations} from '../utils/DatasetUtils';
+import {DatasetService} from '../utils/DatasetService';
 
 export default function DACDatasets(props) {
 
@@ -27,35 +24,13 @@ export default function DACDatasets(props) {
     setFilteredList
   ), [datasets]);
 
-  /**
-   * Async utility to find all datasets a DAC Chairperson
-   * has access to and populate their data use codes and translations
-   * @returns {Promise<Awaited<{translations: ([]|[]|Awaited<unknown>[]), codeList: (string|string)}>[]>}
-   */
-  const populateDacDatasets = async () => {
-    // Find all Chairperson DAC Ids for the user
-    const user = Storage.getCurrentUser();
-    const dacIds = flow(
-      filter(r => r.name === USER_ROLES.chairperson),
-      map('dacId'),
-      filter(id => !isNil(id))
-    )(user.roles);
-    // find all datasets for all DACs the user has access to.
-    const dacDatasets = (await Promise.all(dacIds.map(id => DAC.datasets(id)))).flat();
-    // Enrich datasets with data use translations
-    return await Promise.all(map(async (d) => {
-      const {codeList, translations} = await getDataUseTranslations(d);
-      return Object.assign({}, d, {codeList: codeList, translations: translations});
-    })(dacDatasets));
-  };
-
   useEffect(() => {
     const init = async () => {
       try {
         setIsLoading(true);
-        const enrichedDatasets = await populateDacDatasets();
-        setDatasets(enrichedDatasets);
-        setFilteredList(enrichedDatasets);
+        const dacDatasets = await DatasetService.populateDacDatasets();
+        setDatasets(dacDatasets);
+        setFilteredList(dacDatasets);
         setIsLoading(false);
       } catch (error) {
         Notifications.showError({text: 'Error initializing datasets table'});
